Feat: Changing misc to a dedicated page Refactor: Moving class to dedicated file
Notes: Changing to meet Pheonix's requested change
